fix : 날짜, 특정 교과목으로 등록한 코딩존 수업 정보 리스트 반환 로직 수정 (#344)
* fix : Controller 로직에서 SubjectId 타입을 Integer 로 변경

* fix : Repository 로직에서 SubjectId 타입을 Integer 로 변경

* fix : service 로직에서 SubjectId 타입을 Integer 로 변경

+++ backend/src/main/java/com/icehufs/icebreaker/domain/auth/controller/AdminSubjectController.java
@@ -58,7 +58,7 @@ public class AdminSubjectController {
 
     @GetMapping("/{subjectId}/codingzones")
     public ResponseEntity<ResponseDto<List<CodingZoneClassInfoResponseDto>>> getCodingZoneClassesBySubjectAndDate(
-            @PathVariable Long subjectId,
+            @PathVariable Integer subjectId,
             @RequestParam("date") String date) {
         List<CodingZoneClassInfoResponseDto> result = codingzoneService.findCodingZoneClassesBySubjectAndDate(subjectId, date);
         return ResponseEntity.ok(ResponseDto.success(ResponseMessage.SUCCESS_MAPPING_GET, result));
